[Code Cleanup/Database] Added more error checking to the most common database queries. Removed "or die()" conditions from queries for more graceful handling.

+++ features/cerberusweb.core/api/dao/community_portal.php
@@ -32,7 +32,8 @@ class DAO_CommunityTool extends Cerb_ORMHelper {
 		$sql = sprintf("INSERT INTO community_tool () ".
 			"VALUES ()"
 		);
-		$db->ExecuteMaster($sql) or die(__CLASS__ . '('.__LINE__.')'. ':' . $db->ErrorMsg());
+		if(false == ($db->ExecuteMaster($sql)))
+			return false;
 		$id = $db->LastInsertId();
 		
 		self::update($id, $fields);
@@ -158,6 +159,9 @@ class DAO_CommunityTool extends Cerb_ORMHelper {
 	static private function _createObjectsFromResultSet($rs) {
 		$objects = array();
 		
+		if(!($rs instanceof mysqli_result))
+			return false;
+		
 		while($row = mysqli_fetch_assoc($rs)) {
 			$object = new Model_CommunityTool();
 			$object->id = intval($row['id']);
@@ -196,10 +200,12 @@ class DAO_CommunityTool extends Cerb_ORMHelper {
 			 */
 			
 			$sql = sprintf("DELETE FROM community_tool WHERE id = %d", $id);
-			$db->ExecuteMaster($sql) or die(__CLASS__ . '('.__LINE__.')'. ':' . $db->ErrorMsg());
+			if(false == ($db->ExecuteMaster($sql)))
+				return false;
 			
 			$sql = sprintf("DELETE FROM community_tool_property WHERE tool_code = '%s'", $tool->code);
-			$db->ExecuteMaster($sql) or die(__CLASS__ . '('.__LINE__.')'. ':' . $db->ErrorMsg());
+			if(false == ($db->ExecuteMaster($sql)))
+				return false;
 		}
 	}
 
@@ -280,13 +286,18 @@ class DAO_CommunityTool extends Cerb_ORMHelper {
 			
 		// [TODO] Could push the select logic down a level too
 		if($limit > 0) {
-			$rs = $db->SelectLimit($sql,$limit,$page*$limit) or die(__CLASS__ . '('.__LINE__.')'. ':' . $db->ErrorMsg());
+			if(false == ($rs = $db->SelectLimit($sql,$limit,$page*$limit)))
+				return false;
 		} else {
-			$rs = $db->ExecuteSlave($sql) or die(__CLASS__ . '('.__LINE__.')'. ':' . $db->ErrorMsg());
+			if(false == ($rs = $db->ExecuteSlave($sql)))
+				return false;
 			$total = mysqli_num_rows($rs);
 		}
 		
 		$results = array();
+
+		if(!($rs instanceof mysqli_result))
+			return false;
 		
 		while($row = mysqli_fetch_assoc($rs)) {
 			$object_id = intval($row[SearchFields_CommunityTool::ID]);
@@ -373,6 +384,9 @@ class DAO_CommunityToolProperty extends Cerb_ORMHelper {
 				return false;
 			
 			$props = array();
+			
+			if(!($rs instanceof mysqli_result))
+				return false;
 			
 			while($row = mysqli_fetch_assoc($rs)) {
 				$k = $row['property_key'];
